Moved ctffind metadata operations to mgui. Separated ctffind MGUI from old magicgui.
